backend/requirements.docker.txt
fastapi==0.135.2
uvicorn==0.42.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.2
python-multipart==0.0.22
httpx==0.28.1
httpx-sse==0.4.3
mcp==1.26.0
pydantic==2.12.5
pydantic-settings==2.13.1
starlette==1.0.0
sse-starlette==3.3.3
anyio==4.13.0
sniffio==1.3.1
watchfiles==1.1.1
toonify==1.6.0
PyJWT==2.10.1
bcrypt==4.3.0
